+++ src/main/java/com/personalfinance/backend/service/TransactionService.java
@@ -6,8 +6,8 @@ import com.personalfinance.backend.model.Transaction;
 
 public interface TransactionService {
 
-    public List<Transaction> getAllTransactionsByUserId(long userId);
-    public List<Transaction> getAllTransactionsByUserIdAndMonth(long userId, Integer month);
+    public List<Transaction> getAllTransactionsByUserId(int userId);
+    public List<Transaction> getAllTransactionsByUserIdAndMonth(int userId, Integer month);
     public Transaction addTransaction(Transaction transaction);
     public Transaction updateTransaction(Transaction transaction);
 }
